requirements.txt
pymupdf
pymupdf4llm
transformers 
torch 
pillow
chromadb 
langchain-openai
sentence-transformers
cohere
openai
streamlit
streamlit-extras
langchain-community
tiktoken